requirements.txt
anyio==3.6.2
click==8.1.3
docopt==0.6.2
fastapi==0.95.0
h11==0.14.0
hbmqtt==0.9.6
httptools==0.5.0
idna==3.4
paho-mqtt==1.6.1
passlib==1.7.4
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
transitions==0.9.0
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0
